Moved sampler to data submodule
